requirements.txt
Flask
flask-cors
pandas
numpy
scikit-learn
google-generativeai
gunicorn